pyproject.toml
[project]
name = "gemini-scribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Gal Gilor", email = "gilorgal2010@gmail.com"}
]
readme = "README.md"
packages = [{include = "*", from="gemini_scribe"}]
package-mode = true
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "google-genai (>=1.16.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gcloud-aio-storage (>=9.4.0,<10.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "google-cloud-run (>=0.10.19,<0.11.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
ruff = "^0.11.11"
isort = "^6.0.1"
ipykernel = "^6.29.5"

[tool.ruff]
line-length = 94
fix = true
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "auto"


